통계 페이지 레이아웃 변경 중
