+++ src/assets/css/chart.css
@@ -16,3 +16,7 @@
     font-size: 0.9rem;
 }
 */
+
+.link-box{
+    background-color: bisque;
+}
